refactor: search component 분리
